updated upstream code location and added unit test
